Changed authentication state to be part of the root state
