--- src/app/root-store/authentication/index.ts
@@ -1,6 +1,0 @@
-import * as AuthenticationActions from './actions';
-import * as AuthenticationSelectors from './selectors';
-import { State as AuthenticationState, featureKey as AuthenticationFeatureKey } from './state';
-
-export { AuthenticationStateModule } from './authentication-state.module';
-export { AuthenticationActions, AuthenticationSelectors, AuthenticationState, AuthenticationFeatureKey };
