--- src/app/root-store/authentication/state.ts
@@ -1,15 +1,0 @@
-import { User } from 'src/app/shared/models/user';
-
-export interface State {
-    currentUser: User;
-    isLoading: boolean;
-    error: Error;
-}
-
-export const initialState: State = {
-    currentUser: null,
-    isLoading: false,
-    error: null
-};
-
-export const featureKey = 'authentication';
