+++ src/app/root-store/index.ts
@@ -1,6 +1,6 @@
 import { State as RootState } from './state';
+import * as RootActions from './actions';
+import * as RootSelectors from './selectors';
 
-
-export { RootState };
-export * from './authentication';
+export { RootState, RootActions, RootSelectors };
 export { RootStoreModule } from './root-store.module';
