+++ src/app/root-store/state.ts
@@ -1,5 +1,20 @@
-import { AuthenticationFeatureKey, AuthenticationState } from './authentication';
+import { User } from '../shared/models/user';
 
+
+export const authenticationPropertyKey = 'authentication';
 export interface State {
-    [AuthenticationFeatureKey]: AuthenticationState;
+    [authenticationPropertyKey]: AuthenticationState;
+}
+export const initialState: State = {
+  [authenticationPropertyKey]: {
+    currentUser: null,
+    isLoading: false,
+    error: null
+  }
+};
+
+export interface AuthenticationState {
+  currentUser: User;
+  isLoading: boolean;
+  error: Error;
 }
